Advanced Island prototype

+++ src/Snap.Hutao/Snap.Hutao/Service/Game/Unlocker/Island/IslandEnvironment.cs
@@ -13,7 +13,7 @@ internal struct IslandEnvironment
     public WIN32_ERROR LastError;
     public int Reserved3;
     public float FieldOfView;
-    public float TargetFrameRate;
+    public int TargetFrameRate;
     public bool DisableFog;
     public nuint FunctionOffsetFieldOfView;
     public nuint FunctionOffsetTargetFrameRate;
